Marginal Publication Selection Meta-analysis Model

Density, distribution, and random variate generation for the marginalized distribution of the publication selection meta-analysis model

dmpsnorm(x, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  log = FALSE)

pmpsnorm(q, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  lower.tail = TRUE, log.p = FALSE)

rmpsnorm(n, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta)

Arguments

x, q
vector of quantiles.

theta0
vector of means.

tau
vector of heterogeneity parameters.

sigma
vector of study standard deviations.

alpha
vector of thresholds for publication bias.

eta
vector of publication probabilities, normalized to sum to 1.

log, log.p
logical; If TRUE, probabilities are given as log(p).

lower.tail
logical; If TRUE (default), the probabilities are \(P[X\leq x]\) otherwise, \(P[X\geq x]\).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

These functions assume a normal underlying effect size distribution and one-sided selection on the effects. For the fixed effects publication bias model see psnorm.

References

Hedges, Larry V. "Modeling publication selection effects in meta-analysis." Statistical Science (1992): 246-255.

Moss, Jonas and De Bin, Riccardo. "Modelling publication bias and p-hacking" Forthcoming (2019)

Examples

rmpsnorm(100, theta0 = 0, tau = 0.1, sigma = 0.1, eta = c(1, 0.5, 0.1))

#>   [1] -0.121225029  0.292038621  0.051329059  0.267074928  0.229604485
#>   [6] -0.112495166  0.389674891  0.016715279  0.379370875  0.197575040
#>  [11] -0.240338604  0.386944167  0.196505805  0.365493913  0.229016269
#>  [16]  0.288410367  0.246739719  0.217711025  0.011556901  0.263204349
#>  [21] -0.254595200  0.072775404 -0.134718476  0.113760982  0.204238843
#>  [26]  0.252983766  0.215800314  0.169822649  0.150589094  0.166830509
#>  [31]  0.192398286 -0.089544916  0.008309936  0.100232819 -0.122626609
#>  [36]  0.170238480  0.170761925 -0.128175857  0.008612404  0.244239338
#>  [41]  0.305294280  0.200169526  0.226897460  0.079069527  0.300724670
#>  [46]  0.197138231  0.148797247 -0.015686368  0.241068080  0.204871069
#>  [51]  0.258394521  0.255051063  0.008388778 -0.209985586 -0.193270861
#>  [56] -0.149569421  0.209005366 -0.056079131  0.202547684 -0.086795492
#>  [61] -0.141059753 -0.098712820  0.016312701  0.251651036  0.301561855
#>  [66]  0.169297603  0.234553302 -0.011886072 -0.063264070  0.189107805
#>  [71]  0.250205408  0.195182498  0.180090601  0.256461172  0.253533995
#>  [76]  0.232192367 -0.021270209  0.201089072 -0.118740589  0.255560641
#>  [81]  0.202755941  0.197783266 -0.013809760  0.279555865  0.087335890
#>  [86] -0.003655815  0.094262682 -0.160882739 -0.077601703  0.268393274
#>  [91]  0.249287597  0.025354923  0.057977353  0.350841404 -0.033301322
#>  [96]  0.075078586  0.236010876 -0.032504749  0.113628294 -0.125879104