p-hacking Meta-analysis Model

Density, distribution, and random variate generation for the p-hacking meta- analysis model.

dphnorm(x, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  log = FALSE)

rphnorm(n, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta)

pphnorm(q, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  lower.tail = TRUE, log.p = FALSE)

Arguments

x, q
vector of quantiles.

theta
vector of means.

sigma
vector of study standard deviations.

alpha
vector of thresholds for p-hacking.

eta
vector of p-hacking probabilities, normalized to sum to 1.

log, log.p
logical; If TRUE, probabilities are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

lower.tail
logical; If TRUE (default), the probabilities are \(P[X\leq x]\) otherwise, \(P[X\geq x]\).

Details

These functions assume one-sided selection on the effects. alpha contains the selection thresholds and eta the vector of p-hacking probabilities. theta is the true effect, while sigma is the true standard deviation before selection.

References

Moss, Jonas and De Bin, Riccardo. "Modelling publication bias and p-hacking" Forthcoming (2019)

Examples

rphnorm(100, theta = 0, sigma = 0.1, eta = c(1, 0.5, 0.1))

#>   [1]  0.174205932  0.176655143  0.196473507  0.226094872  0.184113000
#>   [6]  0.185854063  0.216464985  0.254839102  0.224716213  0.214842415
#>  [11]  0.196755965  0.251227290  0.270545577  0.198624785  0.209678990
#>  [16]  0.201624270  0.201949809  0.029587030  0.087856596  0.278783993
#>  [21]  0.210118745  0.258951979  0.025439714  0.194263791  0.228526180
#>  [26]  0.196176376  0.210193506  0.216612647  0.218659565  0.210803909
#>  [31]  0.268621636  0.200306229  0.176627509  0.210475796  0.351884974
#>  [36]  0.224736884  0.197430768  0.221805261  0.234342955  0.200628188
#>  [41]  0.293171925  0.261881178  0.207666167  0.259535971  0.231109266
#>  [46]  0.267565276  0.221085133  0.216940031  0.320009557  0.005678979
#>  [51]  0.307223350  0.269071492  0.258173091  0.230005829  0.294318816
#>  [56]  0.260047857  0.232001345  0.224722088  0.326091427  0.176014759
#>  [61] -0.010265486  0.262700068  0.310237769  0.184469619  0.182036784
#>  [66]  0.208483285  0.211107534  0.206283893  0.257601123  0.217331422
#>  [71]  0.198410651  0.278893128  0.223515635  0.208875385  0.226105671
#>  [76]  0.234890949  0.240887203  0.309715707  0.220692091  0.209225310
#>  [81]  0.212919641  0.195192303  0.175318846  0.252319661  0.182683305
#>  [86]  0.229524045  0.242592311  0.234196485  0.275332148  0.171564221
#>  [91]  0.231750292  0.188484910  0.207504230  0.229530394  0.201099321
#>  [96]  0.240763639  0.333008888  0.374770998  0.246413618  0.201303093