Publication Selection Meta-analysis Model

Density, distribution, quantile, random variate generation, and expectation calculation for the distribution for the publication selection meta-analysis model

dpsnorm(x, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  log = FALSE)

ppsnorm(q, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  lower.tail = TRUE, log.p = FALSE)

rpsnorm(n, theta, sigma, alpha = c(0, 0.025, 0.05, 1), eta)

Arguments

x, q
vector of quantiles.

theta
vector of means.

sigma
vector of study standard deviations.

alpha
vector of thresholds for publication bias.

eta
vector of publication probabilities, normalized to sum to 1.

log, log.p
logical; If TRUE, probabilities are given as log(p).

lower.tail
logical; If TRUE (default), the probabilities are \(P[X\leq x]\) otherwise, \(P[X\geq x]\).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

The effect size distribution for the publication selection model is not normal, but has itself been selected for. These functions assume one-sided selection on the effects. These functions do not assume the existence of an underlying effect size distribution. For these, see mpsnorm.

References

Hedges, Larry V. "Modeling publication selection effects in meta-analysis." Statistical Science (1992): 246-255.

Moss, Jonas and De Bin, Riccardo. "Modelling publication bias and p-hacking" Forthcoming (2019)

Examples

rpsnorm(100, theta = 0, sigma = 0.1, eta = c(1, 0.5, 0.1))

#>   [1]  0.207968641  0.065918555 -0.121244833  0.033671496  0.130009244
#>   [6] -0.026047039 -0.141518387  0.057938809  0.185370843  0.127540210
#>  [11]  0.235569498  0.211486193  0.042020709  0.044536301  0.061318021
#>  [16] -0.164546788 -0.047013513  0.148109003  0.128402652  0.247915003
#>  [21]  0.094197475 -0.010374865  0.028051629  0.220706624  0.070636136
#>  [26]  0.037676946 -0.026436916  0.197321482  0.021584018  0.137683133
#>  [31]  0.036923822  0.148820199  0.203400454  0.167192404 -0.052840324
#>  [36] -0.107863654 -0.023488051  0.164998835  0.018292604  0.214129871
#>  [41] -0.111582792  0.025537509  0.169515843  0.020567685  0.006540921
#>  [46]  0.239710611 -0.011062832 -0.016227053  0.166696442 -0.022059324
#>  [51] -0.064205574  0.376303297 -0.033764603  0.167545905 -0.088317017
#>  [56] -0.027420223 -0.026616879  0.188317742 -0.102457159 -0.086061271
#>  [61]  0.038630438  0.244545879  0.036878757  0.221715863 -0.072469120
#>  [66]  0.024249895  0.030282688  0.070356227 -0.094219937  0.011653752
#>  [71]  0.195322369  0.137925781  0.014958771 -0.056776783  0.179033336
#>  [76]  0.212104990  0.182621556  0.249586620  0.046970481 -0.062113229
#>  [81] -0.145280285 -0.125060060 -0.157491963 -0.149741394  0.235188848
#>  [86] -0.067142028  0.007039665  0.212673054 -0.123002238  0.173482834
#>  [91]  0.157457485 -0.056250751 -0.091506011  0.019631941  0.005961225
#>  [96]  0.214274527 -0.073325992  0.174491351  0.089053264 -0.057483659