Selected Normal Effect Size Distribution

Density, random variate generation, and expectation calculation for the effect size distribution of the one-sided normal publication bias model.

dsnorm(x, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  log = FALSE)

rsnorm(n, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta)

esnorm(theta0, tau, sigma, alpha, eta)

Arguments

x
vector of quantiles.

theta0
vector of means.

tau
vector of heterogeneity parameters.

sigma
vector of study standard deviations.

alpha
vector of thresholds for publication bias.

eta
vector of publication probabilities, normalized to sum to 1.

log
logical; If TRUE, probabilities are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

The effect size distribution for the publication selection model is not normal, but has itself been selected for. These functions assume a normal underlying effect size distribution and one-sided selection on the effects.

References

Hedges, Larry V. "Modeling publication selection effects in meta-analysis." Statistical Science (1992): 246-255.

Moss, Jonas and De Bin, Riccardo. "Modelling publication bias and p-hacking" Forthcoming (2019)

Examples

rsnorm(100, theta0 = 0, tau = 0.1, sigma = 0.1, eta = c(1, 0.5, 0.1))

#>   [1] -0.074259108  0.162280594  0.087490683 -0.092348799  0.174020335
#>   [6]  0.140247432 -0.035586285  0.087800578  0.082502030 -0.070708721
#>  [11]  0.005817145  0.066127757 -0.080027646 -0.128258947 -0.005946397
#>  [16]  0.032642296  0.025374031  0.047001280  0.139502319  0.091698203
#>  [21]  0.072493595  0.110889090  0.081829294 -0.018923658 -0.006528758
#>  [26]  0.121228095  0.053463544  0.096381390  0.228456615  0.027227702
#>  [31]  0.307708815  0.058784853  0.009689439  0.200364435  0.032791961
#>  [36]  0.102046305 -0.035955168  0.174021839  0.041608253  0.248363892
#>  [41]  0.174715026  0.151540668  0.079952564  0.202079836  0.127465313
#>  [46] -0.092042996  0.171166122  0.067116824 -0.128382503  0.026858974
#>  [51]  0.166985957 -0.215574292  0.283926623 -0.042907609  0.030990835
#>  [56] -0.145620804  0.051967261  0.129533072  0.236581164  0.145685797
#>  [61] -0.106011455  0.109196862  0.127654690 -0.025706526  0.104720542
#>  [66]  0.130429027  0.049745036  0.172957275  0.142656559  0.074418187
#>  [71]  0.077324476  0.070059227 -0.058070469  0.095292889  0.246361878
#>  [76]  0.099105105 -0.055378249  0.089319710  0.086303020  0.161549511
#>  [81] -0.297388588  0.054298914  0.031731663  0.110405230  0.223892482
#>  [86]  0.157449189  0.150062561  0.016515865  0.119545735  0.051372190
#>  [91]  0.097011720 -0.013670204 -0.030102955  0.166916698  0.025594315
#>  [96]  0.025008054 -0.035208949  0.009863716 -0.072454559 -0.141335783